projects/files/DataAnalysisTool/api/requirements.txt
flask==3.0.3
psycopg2-binary
